Replit repo change.
basically my replit repo is fucked up because of tfjs making me unable to commit to github. and the more i fiddle around with it it got even more fucked up. so here i am creating new repo just for comitting to github and i must manually move file and shit. and that original repo is hopeless other than just for developing purposes.